chatbot/requirements.txt
# Chatbot Dataiku - Dependencies
streamlit>=1.31.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# Dataiku dependencies (inherited from parent)
dataiku-api-client>=12.0.0
pandas>=2.0.0
